refactoring result
