+++ components/CMakeLists.txt
@@ -12,11 +12,12 @@ else()
     add_subdirectory(log)
     add_subdirectory(logical_plan)
     add_subdirectory(oid)
-    add_subdirectory(ql)
-    add_subdirectory(translator)
-    add_subdirectory(session)
     add_subdirectory(pipeline)
+    add_subdirectory(ql)
+    add_subdirectory(result)
+    add_subdirectory(session)
     add_subdirectory(sql)
+    add_subdirectory(translator)
 endif()
 
 if (DEV_MODE)
