+++ components/result/CMakeLists.txt
@@ -1,7 +1,7 @@
-project(memory_storage)
+project(result)
 
 set(${PROJECT_NAME}_SOURCES
-    memory_storage.cpp
+    result.cpp
 )
 
 add_library(rocketjoe_${PROJECT_NAME}
@@ -14,12 +14,7 @@ set_property(TARGET rocketjoe_${PROJECT_NAME} PROPERTY EXPORT_NAME ${PROJECT_NAM
 
 target_link_libraries(
     rocketjoe_${PROJECT_NAME} PRIVATE
-    rocketjoe::collection
-    rocketjoe::disk
-    rocketjoe::locks
-    rocketjoe::result
-    rocketjoe::session
-    CONAN_PKG::spdlog
+    rocketjoe::ql
 )
 
 target_include_directories(rocketjoe_${PROJECT_NAME} PUBLIC
