+++ integration/cpp/test/test_collection_sql.cpp
@@ -7,6 +7,7 @@ static const database_name_t database_name = "TestDatabase";
 static const collection_name_t collection_name = "TestCollection";
 
 using namespace components;
+using namespace components::result;
 using expressions::compare_type;
 using ql::aggregate::operator_type;
 using key = components::expressions::key_t;
