+++ integration/cpp/test/test_index.cpp
@@ -54,7 +54,7 @@ static const collection_name_t collection_name = "TestCollection";
     do { \
         auto session = duck_charmer::session_id_t(); \
         auto *ql = new components::ql::aggregate_statement{database_name, collection_name}; \
-        auto c = dispatcher->find(session, ql); \
+        auto c = dispatcher->find(session, ql).get<components::cursor::cursor_t*>(); \
         REQUIRE(c->size() == 100); \
         delete c; \
     } while (false)
@@ -66,7 +66,7 @@ static const collection_name_t collection_name = "TestCollection";
         auto expr = components::expressions::make_compare_expression(dispatcher->resource(), COMPARE, key{KEY}, id_par{1}); \
         ql->append(operator_type::match, components::ql::aggregate::make_match(std::move(expr))); \
         ql->add_parameter(id_par{1}, VALUE); \
-        auto c = dispatcher->find(session, ql); \
+        auto c = dispatcher->find(session, ql).get<components::cursor::cursor_t*>(); \
         REQUIRE(c->size() == COUNT); \
         delete c; \
     } while (false)
