+++ integration/cpp/wrapper_dispatcher.hpp
@@ -17,9 +17,7 @@
 #include <components/ql/index.hpp>
 #include <components/ql/aggregate.hpp>
 #include <components/ql/statements.hpp>
-
-#include <services/collection/result.hpp>
-#include <services/memory_storage/result.hpp>
+#include <components/result/result.hpp>
 
 namespace duck_charmer {
 
@@ -32,23 +30,23 @@ namespace duck_charmer {
         wrapper_dispatcher_t(actor_zeta::detail::pmr::memory_resource* , actor_zeta::address_t,log_t &log);
         ~wrapper_dispatcher_t();
         auto load() -> void;
-        auto create_database(session_id_t &session, const database_name_t &database) -> services::memory_storage::result_t;
-        auto drop_database(session_id_t &session, const database_name_t &database) -> services::memory_storage::result_t;
-        auto create_collection(session_id_t &session, const database_name_t &database, const collection_name_t &collection) -> services::memory_storage::result_t;
-        auto drop_collection(session_id_t &session, const database_name_t &database, const collection_name_t &collection) -> services::memory_storage::result_t;
-        auto insert_one(session_id_t &session, const database_name_t &database, const collection_name_t &collection, document_ptr &document) -> result_insert&;
-        auto insert_many(session_id_t &session, const database_name_t &database, const collection_name_t &collection, std::pmr::vector<document_ptr> &documents) -> result_insert&;
-        auto find(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> components::cursor::cursor_t*;
-        auto find_one(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> result_find_one&;
-        auto delete_one(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> result_delete&;
-        auto delete_many(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> result_delete&;
-        auto update_one(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition, document_ptr update, bool upsert) -> result_update&;
-        auto update_many(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition, document_ptr update, bool upsert) -> result_update&;
-        auto size(session_id_t &session, const database_name_t &database, const collection_name_t &collection) -> result_size;
-        auto create_index(session_id_t &session, components::ql::create_index_t index) -> result_create_index;
-        auto drop_index(session_id_t &session, components::ql::drop_index_t drop_index) -> result_drop_index;
-        auto execute_ql(session_id_t& session, components::ql::variant_statement_t& query) -> result_t;
-        auto execute_sql(session_id_t& session, const std::string& query) -> result_t;
+        auto create_database(session_id_t &session, const database_name_t &database) -> components::result::result_t;
+        auto drop_database(session_id_t &session, const database_name_t &database) -> components::result::result_t;
+        auto create_collection(session_id_t &session, const database_name_t &database, const collection_name_t &collection) -> components::result::result_t;
+        auto drop_collection(session_id_t &session, const database_name_t &database, const collection_name_t &collection) -> components::result::result_t;
+        auto insert_one(session_id_t &session, const database_name_t &database, const collection_name_t &collection, document_ptr &document) -> components::result::result_insert&;
+        auto insert_many(session_id_t &session, const database_name_t &database, const collection_name_t &collection, std::pmr::vector<document_ptr> &documents) -> components::result::result_insert&;
+        auto find(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> components::result::result_t;
+        auto find_one(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> components::result::result_t;
+        auto delete_one(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> components::result::result_delete&;
+        auto delete_many(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition) -> components::result::result_delete&;
+        auto update_one(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition, document_ptr update, bool upsert) -> components::result::result_update&;
+        auto update_many(session_id_t &session, components::ql::aggregate_statement_raw_ptr condition, document_ptr update, bool upsert) -> components::result::result_update&;
+        auto size(session_id_t &session, const database_name_t &database, const collection_name_t &collection) -> components::result::result_size;
+        auto create_index(session_id_t &session, components::ql::create_index_t index) -> components::result::result_create_index;
+        auto drop_index(session_id_t &session, components::ql::drop_index_t drop_index) -> components::result::result_drop_index;
+        auto execute_ql(session_id_t& session, components::ql::variant_statement_t& query) -> components::result::result_t;
+        auto execute_sql(session_id_t& session, const std::string& query) -> components::result::result_t;
 
     protected:
 
@@ -58,24 +56,22 @@ namespace duck_charmer {
     private:
         /// async method
         auto load_finish() -> void;
-        auto execute_ql_finish(session_id_t &session, const services::memory_storage::result_t& result) -> void;
-        auto insert_finish(session_id_t &session, result_insert result) -> void;
-        auto find_finish(session_id_t &session, components::cursor::cursor_t *cursor) -> void;
-        auto find_one_finish(session_id_t &session, result_find_one result) -> void;
-        auto delete_finish(session_id_t &session, result_delete result) -> void;
-        auto update_finish(session_id_t &session, result_update result) -> void;
-        auto size_finish(session_id_t &session, result_size result) -> void;
-        auto create_index_finish(session_id_t &session, result_create_index result) -> void;
-        auto drop_index_finish(session_id_t &session, result_drop_index result) -> void;
+        auto execute_ql_finish(session_id_t &session, const components::result::result_t& result) -> void;
+        auto insert_finish(session_id_t &session, components::result::result_insert result) -> void;
+        auto delete_finish(session_id_t &session, components::result::result_delete result) -> void;
+        auto update_finish(session_id_t &session, components::result::result_update result) -> void;
+        auto size_finish(session_id_t &session, components::result::result_size result) -> void;
+        auto create_index_finish(session_id_t &session, components::result::result_create_index result) -> void;
+        auto drop_index_finish(session_id_t &session, components::result::result_drop_index result) -> void;
 
         void init();
         void wait();
         void notify();
 
         template <typename Tres, typename Tql>
-        auto send_ql(session_id_t &session, Tql& ql, std::string_view title, uint64_t handle) -> result_t;
+        auto send_ql(session_id_t &session, Tql& ql, std::string_view title, uint64_t handle) -> components::result::result_t;
 
-        auto send_ql_new(session_id_t &session, components::ql::ql_statement_t* ql) -> services::memory_storage::result_t;
+        auto send_ql_new(session_id_t &session, components::ql::ql_statement_t* ql) -> components::result::result_t;
 
         actor_zeta::address_t manager_dispatcher_;
         log_t log_;
@@ -85,17 +81,15 @@ namespace duck_charmer {
         std::condition_variable cv_;
         session_id_t input_session_;
         std::variant<
-            null_result,
-            result_insert,
-            components::cursor::cursor_t*,
-            result_find_one,
-            result_size,
-            result_delete,
-            result_update,
-            result_drop_collection,
-            result_create_index,
-            result_drop_index,
-            services::memory_storage::result_t>
+            components::result::empty_result_t,
+            components::result::result_insert,
+            components::result::result_size,
+            components::result::result_delete,
+            components::result::result_update,
+            components::result::result_drop_collection,
+            components::result::result_create_index,
+            components::result::result_drop_index,
+            components::result::result_t>
             intermediate_store_;
     };
 } // namespace python
