+++ integration/python/wrapper_result.hpp
@@ -3,7 +3,7 @@
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
 #include <boost/smart_ptr/intrusive_ref_counter.hpp>
-#include "services/collection/result.hpp"
+#include <components/result/result.hpp>
 #include "wrapper_cursor.hpp"
 
 namespace py = pybind11;
@@ -13,7 +13,7 @@ namespace duck_charmer {
 class PYBIND11_EXPORT wrapper_result_delete final : public boost::intrusive_ref_counter<wrapper_result_delete> {
 public:
     wrapper_result_delete() = default;
-    wrapper_result_delete(const result_delete &src);
+    wrapper_result_delete(const components::result::result_delete &src);
 
     const py::list &raw_result() const;
     std::size_t deleted_count() const;
@@ -26,7 +26,7 @@ private:
 class PYBIND11_EXPORT wrapper_result_update final : public boost::intrusive_ref_counter<wrapper_result_update> {
 public:
     wrapper_result_update() = default;
-    wrapper_result_update(const result_update &src);
+    wrapper_result_update(const components::result::result_update &src);
 
     py::list raw_result() const;
     std::size_t matched_count() const;
@@ -34,14 +34,14 @@ public:
     py::object upserted_id() const;
 
 private:
-    result_update result;
+    components::result::result_update result;
 };
 
 
 class PYBIND11_EXPORT wrapper_result final : public boost::intrusive_ref_counter<wrapper_result> {
 public:
     wrapper_result();
-    wrapper_result(const duck_charmer::session_id_t& session, const result_t &result);
+    wrapper_result(const duck_charmer::session_id_t& session, const components::result::result_t &result);
 
     std::size_t inserted_count() const;
     std::size_t modified_count() const;
@@ -50,7 +50,7 @@ public:
 
 private:
     duck_charmer::session_id_t session_;
-    result_t result_;
+    components::result::result_t result_;
 };
 
 }
