+++ services/collection/create_index.cpp
@@ -10,6 +10,8 @@ using components::ql::index_type;
 using components::index::make_index;
 using components::index::single_field_index_t;
 
+using namespace components::result;
+
 namespace services::collection {
 
     std::string name_index_type(index_type type) {
@@ -118,7 +120,7 @@ namespace services::collection {
             }
         }
         sessions::remove(sessions_, session);
-        actor_zeta::send(suspend_plan.client, address(), handler_id(route::find_finish), session, res.first->second.get());
+        //actor_zeta::send(suspend_plan.client, address(), handler_id(route::find_finish), session, res.first->second.get());
     }
 
 } // namespace services::collection
