+++ services/collection/route.hpp
@@ -7,8 +7,6 @@ namespace services::collection {
     enum class route : uint64_t {
         create_documents,
         insert_documents,
-        find,
-        find_one,
         delete_documents,
         update_documents,
         size,
@@ -19,8 +17,6 @@ namespace services::collection {
 
         create_documents_finish,
         insert_finish,
-        find_finish,
-        find_one_finish,
         delete_finish,
         update_finish,
         size_finish,
